Read処理のServiceとController部分を実装 ・パッケージの追加とクラスの移動 ・サービスに抽出処理を追加
